+++ CHANGELOG.md
@@ -1,14 +1,24 @@
+## 1.1.0
+
+* Added option to specify the value to be displayed or selectable for the time picker.
+* Execute change callback if initial date is outside minimum or maximum range
+* [Bug fix] Fixed problem with imps and fields not entering values correctly
+
 ## 1.0.3
+
 * Allows minimum and maximum dates to be specified
 * Added `openPicker` that can be displayed without arguments when open.
 * [Bug fix] Fixed date misalignment issue with calendar paging
 
 ## 1.0.2
+
 * Changed to allow specification of language data to be displayed
 * [Bug fix] Fixed problem with date buttons not becoming selected
 
 ## 1.0.1
+
 * Update README
 
 ## 1.0.0
+
 * Release board_datetime_picker